+++ src/App.svelte
@@ -2,6 +2,7 @@
     import { afterUpdate, onDestroy, tick } from 'svelte'
 
     const internalPrefix = '__gg__'
+    const nodeRadius = 5
     let svgWidth = 100
     let svgHeight = 100
 
@@ -37,6 +38,16 @@
         placed = false
     }
 
+    // _connectors is an array of of objects in the form
+    // {
+    //   from: string
+    //   to: string
+    //   color?: string
+    //   x1?: number
+    //   y1?: number
+    //   x2?: number
+    //   y2?: number
+    // }
     let _connectors = []
     $: {
         _connectors = JSON.parse(connectors)
@@ -103,6 +114,55 @@
             if (node.box) {
                 width = Math.max(node.box.x2, width)
                 height = Math.max(node.box.y2, height)
+
+                const sideTop = node.box.y1 + nodeRadius
+                const sideHeight = node.box.y2 - node.box.y1 - 2 * nodeRadius
+
+                // Spread out all connectors from this node
+                // to a node to the right on the right side of this node
+                const connFromRight = _connectors
+                    .filter(conn => conn.from === node.id)
+                    .filter(conn => {
+                        const toNode = nodeMap.get(conn.to)
+                        return toNode && toNode.box.x1 > node.box.x2
+                    })
+                    .sort((a, b) => {
+                        // Sort connectors based on the y-position
+                        // of the node they are connecting to
+                        const nodeA = nodeMap.get(a.to)
+                        const nodeB = nodeMap.get(b.to)
+                        return nodeA.box.y1 - nodeB.box.y1
+                    })
+
+                const sideSectionRight = sideHeight / (connFromRight.length + 1)
+
+                connFromRight.forEach((conn, i) => {
+                    conn.x1 = node.box.x2
+                    conn.y1 = sideTop + sideSectionRight * (i + 1)
+                })
+
+                // Spread out all connectors to this node
+                // from a node to the left on the left side of this node
+                const connToLeft = _connectors
+                    .filter(conn => conn.to === node.id)
+                    .filter(conn => {
+                        const fromNode = nodeMap.get(conn.from)
+                        return fromNode && fromNode.box.x2 < node.box.x1
+                    })
+                    .sort((a, b) => {
+                        // Sort connectors based on the y-position
+                        // of the node they are connecting from
+                        const nodeA = nodeMap.get(a.from)
+                        const nodeB = nodeMap.get(b.from)
+                        return nodeA.box.y1 - nodeB.box.y1
+                    })
+
+                const sideSectionLeft = sideHeight / (connToLeft.length + 1)
+
+                connToLeft.forEach((conn, i) => {
+                    conn.x2 = node.box.x1 - 4
+                    conn.y2 = sideTop + sideSectionLeft * (i + 1)
+                })
             }
         }
 
@@ -110,16 +170,22 @@
         svgHeight = height
 
         _connectors = _connectors.map(conn => {
-            const box1 = nodeMap.get(conn.from).box
-            const box2 = nodeMap.get(conn.to).box
+            const node1 = nodeMap.get(conn.from)
+            const node2 = nodeMap.get(conn.to)
+            if (!node1) {
+                console.log('No such node ' + conn.from)
+                return {...conn, x1: undefined}
+            }
+            if (!node2) {
+                console.log('No such node ' + conn.to)
+                return {...conn, x1: undefined}
+            }
+            const box1 = node1.box
+            const box2 = node2.box
             if (box1 && box2) {
                 if (box1.x2 < box2.x1) {
-                    // box1 left of box2
-                    const x1 = box1.x2
-                    const x2 = box2.x1 - 4
-                    const y1 = middle(box1.y1, box1.y2)
-                    const y2 = middle(box2.y1, box2.y2)
-                    return { ...conn, x1, y1, x2, y2 }
+                    // Already placed
+                    return conn
                 } else if (middle(box2.x1, box2.x2) < box1.x1) {
                     console.log('from cannot be to the right of to')
                 } else {